LayoutTests/fast/forms/option-value-and-label-changed-by-js.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<select>
<option id="o1">  text  </option>
<option id="o2" value="value">  text  </option>
<option id="o3" label="label">  text  </option>
<option id="o4" value="value" label="label">text</option>
</select>

<script>
description('Test for .value and .label of OPTION element');

var o1 = document.getElementById('o1');
o1.label = "newLabel"
o1.value = "newValue"
shouldBe('o1.value', '"newValue"');
shouldBe('o1.label', '"newLabel"');

var o2 = document.getElementById('o2');
o2.label = "newLabel"
o2.value = "newValue"
shouldBe('o2.value', '"newValue"');
shouldBe('o2.label', '"newLabel"');

var o3 = document.getElementById('o3');
o3.label = "newLabel"
o3.value = "newValue"
shouldBe('o3.value', '"newValue"');
shouldBe('o3.label', '"newLabel"');

var o4 = document.getElementById('o4');
o4.label = "newLabel"
o4.value = "newValue"
shouldBe('o4.value', '"newValue"');
shouldBe('o4.label', '"newLabel"');
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
